move to internal/gitserve
