+++ internal/gitserve/serve_test.go
@@ -1,4 +1,4 @@
-package main
+package gitserve
 
 import (
 	"encoding/json"
